Ajout de fonction bis

+++ include/tree.h
@@ -1,20 +1,18 @@
 #ifndef _TREE_H_
 #define _TREE_H_
 
-#include <stddef.h>
-
 #include "tokens.h"
 
 typedef struct tree {
-  size_t count;
-  size_t size;
+  int count;
+  int size;
   struct tree **child;
   Token token;
 } *Tree;
 
 Tree newTree();
 void setValue(Tree t, Token value);
-void setNbChild(Tree t, size_t nb);
+void setNbChild(Tree t, int nb);
 void addChild(Tree parent, Tree child);
 void deleteTree(Tree t);
